+++ Game Of Life - App/Classes/Preferences.swift
@@ -32,7 +32,7 @@ import Cocoa
     @objc public dynamic var drawAsSquares:    Bool    = false
     @objc public dynamic var preserveGridSize: Bool    = true
     @objc public dynamic var speed:            UInt    = 10
-    @objc public dynamic var cellSize:         CGFloat = 10
+    @objc public dynamic var cellSize:         UInt    = 10
     
     @objc private dynamic var color1RGB: UInt = 0xFFFFFF
     @objc private dynamic var color2RGB: UInt = 0xFFFFFF
